Create Player
Create entities script which will contain a player class. Player class will load a defined image and assign a size to it.

In main a for loop will get the Entities layer from the tiled map. The loop will run logic to look for the Player object and check to see if the pos matches the defined player_start_position (which currently is defined as 'house'). If the logic finds it, it will render the sprite.