Camera Follow Player
In main store the Player method call in a variable named self.player. Pass the players center position to the draw method call.

In groups, create a vector and store it in a offset variable. Update the draw method to accept the players center position, offset the x and y of the camera by the opposite of the players location in the window.

Use surface.blit to draw the player sprite on another sprite every draw call.